CrimeMapBE/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.0
black==19.10b0
click==7.1.2
Django==3.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-gis==0.15
lxml==4.5.0
pathspec==0.8.0
postgis==1.0.4
psycopg2==2.8.5
pytz==2020.1
regex==2020.5.14
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
